moved SpliceRegion type to shared
