+++ shared/types/SpliceRegion.ts
