+++ tsconfig.json
@@ -22,6 +22,6 @@
       ]
     },
   },
-  "include": ["src"],
+  "include": ["src", "shared"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
